requirements.txt
Brotli @ file:///C:/b/abs_3d36mno480/croot/brotli-split_1714483178642/work
certifi @ file:///C:/b/abs_1fw_exq1si/croot/certifi_1725551736618/work/certifi
chardet==5.2.0
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
colorama==0.4.6
decorator==4.4.2
exceptiongroup==1.2.2
future==1.0.0
huepy==1.2.1
idna @ file:///C:/b/abs_aad84bnnw5/croot/idna_1714398896795/work
imageio==2.36.0
imageio-ffmpeg==0.5.1
iniconfig==2.0.0
instabot==0.117.0
mock==5.1.0
moviepy==1.0.3
numpy==2.0.2
packaging==24.1
pillow==11.0.0
pluggy==1.5.0
proglog==0.1.10
PySocks @ file:///C:/ci/pysocks_1605307512533/work
pytest==8.3.3
pytz==2024.2
PyYAML==6.0.2
requests @ file:///C:/b/abs_9frifg92q2/croot/requests_1721410901096/work
requests-toolbelt==1.0.0
responses==0.25.3
schedule==1.2.2
six==1.16.0
tomli==2.0.2
tqdm==4.66.6
urllib3 @ file:///C:/b/abs_9a_f8h_bn2/croot/urllib3_1727769836930/work
win-inet-pton @ file:///C:/ci/win_inet_pton_1605306162074/work
